Make the bootstrapping process work again after moving tools to the @bazel_tools repository.
The fix is to always compile tools when the Bazel binary is compiled and add the bootstrap arguments to bootstrap_test. This stil litters the git client with output files in random places, which is a bit difficult to fix, since //src:bazel expects these files to be there so that it can embed them into the output binary.

I didn't notice this issue because tools/jdk/*_deploy.jar is in the .gitignore file, thus, when "git status" reported nothing, I asumed I ran the build in a clean client.

--
MOS_MIGRATED_REVID=105206124