Automated [] rollback of [] + merge with []
*** Reason for rollback ***

Original CL uncovered a depot issue which was fixed in []. I verified it was the only such issue (see []).

*** Original change description ***

Rollback of commit 145aa9594111a32ac9e8b1b030bcf04132d63c5b.

*** Reason for rollback ***

Rolling back until [] is submitted and we have verified that there are no other breakages

*** Original change description ***

Handle the case of infinite symlink expansion where a path in a symlink chain is a symlink to an ancestor of a path in the chain.

--
MOS_MIGRATED_REVID=105251788