Fix AbstractQueueVisitor synchronization, comments, and field names
This fixes the following synchronization issue with
AbstractQueueVisitor's jobsMustBeStoppedField: it was read in
awaitTermination in a block synchronized on zeroRemainingTasks, but
in markToStopAllJobsIfNeeded it was read and written in a block
synchronized on the AQV instance. Now, it is always read or written
in a block synchronized on zeroRemainingTasks, because it is used in
the condition represented by that object.

This also thoroughly cleans up obsolete and irregular documentation in
the class.

--
MOS_MIGRATED_REVID=106849236