Skylark: Updated the locations of more error messages.
In the presence of a Skylark macro some error messages used to show the location of a bzl file for errors that actually happened in a BUILD file.

For the "crosses boundary of subpackage" case, we decided to always show the location of the BUILD file. This solution is not perfect since it misses some scenarios where the illegal label was actually specified in a bzl file, such as

def macro_skylark_rule(name, srcs=[]):
  skylark_rule(name = name, srcs = srcs + ['sub/package/illegal.h'])

However, the current design in regards to attribute errors does not allow us to provide two locations (BUILD and bzl), nor can we make a case-by-case choice whether we return the location in the BUILD file or in the bzl file. Consequently, we still get a wrong location in some cases. However, these cases should be rare.

--
MOS_MIGRATED_REVID=107044681