Provide a looser concept of BuildConfiguration "equality" to account for dynamic configurations.
In short, as a configuration gets passed down its dependency tree it's
likely to lose fragments. For example, when a java_library depends on a cc_library,
the former has both JavaConfiguration and CppConfiguration while the latter
only has CppConfiguration.

It's still the "same" configuration in both cases, it's just that it got trimmed.

This cl provides a looser concept of BuildConfiguration equality that
accounts for this.

--
MOS_MIGRATED_REVID=107598262