Rollback of commit 76ca930acc15bebbbb7e1b19a1f2001a8627b3be.
*** Reason for rollback ***

Broke bazel

*** Original change description ***

Ensure that the plist inside an .ipa bundle produced by blaze and the adjacent plist read by xcode are identical.

To do this, we use the output of plmerge as the single plist for the bundle.  Automatic entries and variable substitutions are both computed in blaze and passed into plmerge.  The output of plmerge is passed into bundlemerge to be placed directly into the final bundle.

--
MOS_MIGRATED_REVID=110792182