Use label syntax that is immune to MSYS path mangling in bootstrap scripts.
On Windows, MSYS will mangle all arguments that resemble Unix paths
when executing (exec*()) non-msys executables (in an attempt to convert
them to Windows paths). This affects ``//src:bazel`` (it becomes
``/src:bazel``) but not ``src:bazel``. This CL converts to the latter in
bootstrapping shell scripts to work around this issue.

--
MOS_MIGRATED_REVID=113349821