Contractually document BuildConfiguration.Fragment as immutable and strengthen the immutability interfaces of existing implementations.
Eventually we want all implementations to comply, but right now CppConfiguration
is a glaring exception due to FDO/LIPO support.

We don't want more exceptions to arise.

This is prep work for pre-trimming ConfigurationFragment.key's BuildOptions input
to just the options needed by the fragment. That implies fragments can be shared across configurations, so that needs to be safe.

--
MOS_MIGRATED_REVID=113408041