Rollback of commit 6ecf87234bf03f52443810827a44cb95ab7754be.
*** Reason for rollback ***

Breaks targets. See []

Manually verified this unbreaks the target mentioned in the bug.

*** Original change description ***

Do not compile source files mentioned in 'hdrs'.

Currently for a library:
  cc_library(name='a', hdrs=['a.cc'])
we compile a.cc into a.pic.o and link that into its reverse dependencies.

With this change, a .cc file in hdrs will be treated like a .inc file or
a file in textual_hdrs.

--
MOS_MIGRATED_REVID=113568452