Use hard links to create empty files.
This speeds up my test of running a bunch of genrules with 3000 inputs
by 2x on XFS. It might speed up ext4 a bit, but it definitely doesn't
make it slower. XFS and ext4 are about the same speed with this. Test at
<https://gist.github.com/bsilver8192/10527a862ce16bb7f79a>.

Refs #852

RELNOTES: Improved sandbox performance on XFS filesystems.

--
Change-Id: Ifabc0442f9f919d50cb5fe80a6445890c6d7f03e
Reviewed-on: https://bazel-review.googlesource.com/#/c/2890/1
MOS_MIGRATED_REVID=114693725