Split out Java proto compilation logic to a helper script (take 2).
This is just for clarity and should be a no-op.  Maintaining long shell
command lines embedded in genrules is not pretty.

This change is a combination of commit 5978125f5955c9d56f74230101e09a0401de8130
and commit f24ba706d3537c3b12edf0bc1f6abe5fdae6d5dd with additional fixes.
In particular I have removed unnecessary properties from all rule
attributes, have made the rule respect any runfiles, and have reinstanted
the explicit dependency on the JDK.  The latter is what fixes the breakage
we observed in #938, though it seems clear that we are missing a
dependency somewhere.

I have verified that this now works by running this change on a clean
Ubuntu 15.10 VM.

--
MOS_MIGRATED_REVID=115081830