Make project generated by setup-intellij.sh error-free.
Fix three issues, one cosmetic, with setup-intellij.sh: running bazel
query "deps(${target}, 1)" on generated java files with a visibility
specification was returning two targets, one of which was the
visibility. Filtering for rules fixes that.

Bazel uses the javac langtools in third_party/ in preference to the
SDK's, but that wasn't reflected in the order of dependencies in the
module. This requires a small hack, since the SDK should appear before
some other jars in the dependencies.

Finally, remove some apparently dead code related to ijar that was just
printing errors when the script was run.

--
Change-Id: Iba5ae5a46c720282d99f193cd098ea14cc5eef52
Reviewed-on: https://bazel-review.googlesource.com/2990
MOS_MIGRATED_REVID=115246698