Restore allow_files and single_file attributes.
//src/test/shell/bazel:bazel_example_test broke because of commit c6e136f47fe6cef9c9958cd82d4db13eb6c64eae,
which removed the allow_files and single_file properties of a bunch of
attributes in the belief that they were stale.  However, //third_party:protoc
is in fact a filegroup and thus requires single_file.

I still think some of them are unnecessary... but because I do not yet know
all possible cases in which these are involved, and to prevent further
breakage, let's restore them all just in case.  I may reevaluate this at a
later stage to clean them up.

--
MOS_MIGRATED_REVID=115297686