experimental UI: only refresh progress bar if needed
The handle(Event event) method of the ExperimentalEventHandler does
not produce UI changes for every event observed. Still, it used to
clear and re-add the progress bar unconditionally. Move those operations
into the individual cases that do produce output, so that in the ignored
events nothing is sent to the terminal.

--
Change-Id: I31fd8df0ee2feeb53b6686fe7350848be2a53b6a
Reviewed-on: https://bazel-review.googlesource.com/#/c/3025
MOS_MIGRATED_REVID=115656142