Add wrapper class around AnsiTerminalWriter that breaks lines
In order to update Bazel's progress bar, the old one has to be removed
first; this requires knowledge about the number of lines it currently
uses. For small terminals, this requires us to take line breaks into
account. While this normally works well, there are situations where our
believe about the terminal width is smaller that the actual width of the
terminal, causing the deletion of too many lines. This wrapper class provides
a solution for this case by explicitly breaking lines at the given width.

--
Change-Id: I9d33e389730568ab8c15ee082594de9b35abb71c
Reviewed-on: https://bazel-review.googlesource.com/#/c/3023
MOS_MIGRATED_REVID=115827354