experimental UI: don't update progress bar after end of build
After completion of the build, there is no need to update the
progress bar any more. Its last line can just move to the scrollback
buffer. The advantage of doing so, is that we then can pass through
STDOUT and STDERR events without additional interference, like line
ending to be sure to know where we position the progress bar and
additional control characters to position the progress bar itself.
In particular, the preformated status report of the tests that were
run will be displayed properly. This makes the experimental UI also
minimally usable for tests.

--
Change-Id: Idaa389b93fc8c9c46c0930f66b4f69b16c3d2e0b
Reviewed-on: https://bazel-review.googlesource.com/#/c/3046
MOS_MIGRATED_REVID=116350626