Expose runfiles symlink functionality in Skylark
The Skylark rule context object has a runfiles method.  This adds two
optional parameters to that method, "symlinks" and "root_symlinks",
that expose functionality from the underlying Runfiles java class.
With this functionality, one can construct links in the runfiles tree
where the source and destination of the link have different names
and/or relative directories.  This might be useful for things like
AppEngine rules where a file in a subdirectory of the source tree
needs to appear in the root directory of the runfiles tree.

If either new parameter is used, the runfiles is subject to stricter
validity checking.  This checking propagates to other runfiles that
depend on it.

RELNOTES: Added "root_symlinks" and "symlinks" parameters to Skylark
runfiles() method.

--
MOS_MIGRATED_REVID=116879064