Make bootstrap process wrapper honor '-' as stdout/stderr
For the process wrapper the value '-' has a special meaning
as file name to which stdout/stderr are to be redirected: do
not redirect. However, the simple shell script that serves
as process wrapper during the bootstrap phase was not aware
of this special meaning, resulting in loss of useful process
output, which is especially annoying during bootstrap. Fix this.

--
Change-Id: Ifcf84e9000d74dafc69b675f192c1fc1cce484e8
Reviewed-on: https://bazel-review.googlesource.com/#/c/3081
MOS_MIGRATED_REVID=117338558