Remove magic terminal width 97
Bazel expects to be told the width of the terminal; during
bootstrap, we do not have a good mechanism of determining it,
so we have to resort to a hard-coded value. However, 97 is a
not very well motivated value (and actually not a very useful
value either) which is hard to guess that this is a default value.
Resorting to Bazel's built-in default of 80 is only slightly
more useful, but, at least, it is easily recognizable as a default
value, as 80 is the width of the famous VT100, the standard punch
card, etc.

--
Change-Id: I62403ca37ee74a090067a5a4248028e3d624b7c6
Reviewed-on: https://bazel-review.googlesource.com/#/c/3082
MOS_MIGRATED_REVID=117346018