Only break lines if cursor motion is in use
Manual breaking lines was introduced in b8b46d61 to ensure Bazel knows
the number of lines written, so the correct number of lines can be
erased when the progress bar is updated. As this is only an issue if curses
are in use, that wrapping as only added to the FancyTerminalEventHandler.
However, Bazel distinguishes between two uses of curses
- curses to add color to the output, and
- curses to control the cursor and erase outdated progress bars.
If the former is used, but not the latter, it is not necessary to wrap
lines; in fact, not doing so is more backwards compatible while still
keeping the line-eating bug fixed. So don't break lines in this case.

--
Change-Id: Ieef8ab9410a071c35609a395f45cd56653115426
Reviewed-on: https://bazel-review.googlesource.com/#/c/3170
MOS_MIGRATED_REVID=118587980