cc_configure.bzl: fix path to static crosstool
Static crosstool where copied in the wrong path, leading
to not found package and breaking the Windows build.

Also the BUILD.bazel file contains now a reference
to @cc_configure// package, so we need another copy.

Tested: manually patched the change on the Windows slave and
        ran ./compile.sh

Fixes #1101.

--
MOS_MIGRATED_REVID=118674922