cpp: remove unnecessary visibility attribute from blaze_util target
Commit 46a9c1f04f4d: ("Some minor updates to the BUILD files.") made the
default visibility for the whole package public, and thus setting it
again in blaze_util target is rendundat.

While at it remove blaze_util_platform.h from srcs attribute of client
target, as it is already in hdrs list of blaze_util target and client
already depends on it, and thus it is not necessary to list it again
here.

--
Change-Id: I995c3cd6e2f63384acef4b24b9a082e7cb59331a
Reviewed-on: https://bazel-review.googlesource.com/#/c/3150/1
MOS_MIGRATED_REVID=119041634