Keep output of experimental UI short if curses cannot be used
As the experimental UI should be usable in all situations, we also have
to care about the situation where updating the status bar in place is not
possible. In this case, we have to make sure we do not litter the output
to much. To achieve this,
- we use the short one-line version of the status bar, and
- reduce the frequency for time-based updates.
Unfortunately, this means that we have to further complicate the timing mechanisms
for updating the progress bar. We now have 3 time intervals in place:
- a short one, after which we update the progress bar, if we dropped an update
  due to too frequent events,
- an intermediate one, describing the frequency of status bar updates due to time passing, and
- a long one for purely time-based updates of the progress bar if we cannot update it in place.

--
Change-Id: I5d59ba174c4d290b07181620e238362a8d21a6eb
Reviewed-on: https://bazel-review.googlesource.com/#/c/3295
MOS_MIGRATED_REVID=119527089