Make the genproto rules of Bazel usable from other repositories.
This is mainly so that Dash can be fixed. It's not the optimal solution because then Dash will pull in the whole Bazel repository for a tiny little part of it, though. The optimal solution would be to make Bazel use WORKSPACE dependencies and then factor it out to a separate repository, but this will be good enough to unbreak Dash for the time being.

Progress toward fixing https://github.com/bazelbuild/dash/issues/12 .

--
MOS_MIGRATED_REVID=119633183